Refactor to fix travis
