+++ server/controller/index.js
@@ -1,0 +1,3 @@
+export { default as UserController } from './UserController';
+export { default as ReviewController } from './ReviewController';
+export { default as BusinessController } from './BusinessController';
